Move all of the refresh logic into SessionRefreshModal
